edits

+++ src/app/Components/main/UpdateDoctorInfo/update-doctor-certificates/update-doctor-certificates.component.ts
@@ -144,13 +144,43 @@ export class UpdateDoctorCertificatesComponent implements OnInit {
 
   //#region Delete Certificate
     DeleteCertificate(id:number){
-      this.certificateService.DeleteCertificate('en',id).subscribe((res)=>{
-        this.certificateService.GetDoctorCertificate('en').subscribe((res)=>{
-          this.submittedCertificate= res as CertificateResponse
-          console.log(this.submittedCertificate)}
-        )},
-        (err)=>{console.log(err)})
-      }
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Yes, delete it!'
+      })
+      .then((result) => {
+
+        if (result.isConfirmed) {
+          this.certificateService.DeleteCertificate('en',id).subscribe((res)=>{
+            this.certificateService.GetDoctorCertificate('en').subscribe((res)=>{
+              this.submittedCertificate= res as CertificateResponse
+              console.log(this.submittedCertificate)}
+            )
+            Swal.fire(
+              'Deleted!',
+              'Your file has been deleted.',
+              'success'
+            )
+        },
+        (err)=>{
+          console.log(err)
+          Swal.fire("An error occur");
+        })
+         
+        } else {
+          Swal.fire(
+            'Cancelled',
+            'Your imaginary file is safe :)',
+            'error'
+          );
+        }
+      }); 
+    }
   //#endregion
 
 
@@ -170,21 +200,20 @@ export class UpdateDoctorCertificatesComponent implements OnInit {
 
 
   Edit(id:number){
-    
     this.editableCertificate= this.submittedCertificate.Data.find((item)=>item.Id==id) as Certificate
     this.sendButton=true
   }
-  
+
   SaveCertificate(){
     const formData = new FormData();
 
-    formData.append('CertificateId',+this.CertificateForm.get('Id').value as unknown as Blob)
-    formData.append('Title', this.CertificateForm.get('Title').value )
-    formData.append('Description',this.CertificateForm.get('Description').value)
-    formData.append('TitleAr',this.CertificateForm.get('TitleAr').value)
-    formData.append('DescriptionAr',this.CertificateForm.get('DescriptionAr').value)
-    formData.append('Year', +this.CertificateForm.get('Year').value as unknown as Blob)
-    formData.append('certificateImage',this.CertificateForm.get('CertificateUrl').value)
+    formData.append('CertificateId',+this.editableCertificate.Id as unknown as Blob)
+    formData.append('Title',this.editableCertificate.Title)
+    formData.append('Description',this.editableCertificate.Description)
+    formData.append('TitleAr',this.editableCertificate.TitleAr)
+    formData.append('DescriptionAr',this.editableCertificate.DescriptionAr)
+    formData.append('Year', +this.editableCertificate.Year as unknown as Blob)
+    formData.append('certificateImage',this.editableCertificate.CertificateUrl)
 
     this.UpdateCertificate('en',formData)
   }
